backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
python-dotenv==1.0.0
openai==1.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0

